requirements.txt
opencv-python
numpy
scipy
matplotlib
pillow
pandas
